Выключил отображение надписи с кнопкой регистрации "Зарегистрируйтесь чтобы начать добавлять рецепты" если пользователь авторизован; Исправил баг (мною же созданный) когда вылетала ошибка при попытке вызова открытия модального окна откуда либо кроме Header'a или Main'а. Для этого я удалил все Header'ы из страниц и пихнул его в App.js. Таким образом Header отображается на любой странице. После чего перенес 3 useState'а и функцию открытия модального окна из Main.js в App.js, чтобы пробросить их в Header;;; Чтобы пропадала надпись с кнопкой перенес useState modalIsLoggedIn из Header в App, чтобы пробросить его в Main, а потом в RecipesMain, в котором он используется.
