+++ src/components/RecipesMain.js
@@ -4,7 +4,7 @@ import RecipesMainCircle from './RecipesMainCircle';
 import { Link } from "react-router-dom";
 import ModalWindow from './ModalWindow.js';
 
-const RecipesMain = ({openModal}) => {
+const RecipesMain = ({isLoggedIn, setIsLoggedIn, openModal}) => {
   return (
     <div className='RecipesMain'>
       <div className="Circle">
@@ -37,15 +37,27 @@ const RecipesMain = ({openModal}) => {
           <Link to="/AllRecipes" className='Circle__btn-link'>Посмотреть<br/> рецепты</Link>
         </button>
       </div>
-      <div className="SignUp">
-        <div className="SignUp__top">
-          <button className='SignUp__top-link' onClick={() => openModal('register')} >Зарегистрируйтесь</button>,<br/>
-          чтобы начать добавлять<br/>
-          рецепты
+      {isLoggedIn ? (
+        <div className="SignUp">
+          <div className="SignUp__topBlank">
+            {/* <button className='SignUp__top-link' onClick={() => openModal('register')} >Зарегистрируйтесь</button>,<br/> */}
+            {/* чтобы начать добавлять<br/> */}
+            {/* рецепты */}
+          </div>
+          <img src="./img/girlMain.png" alt="" className='SignUp__girl'/>
+          <div className="SignUp__bottom"></div>
         </div>
-        <img src="./img/girlMain.png" alt="" className='SignUp__girl'/>
-        <div className="SignUp__bottom"></div>
-      </div>
+        ) : (
+          <div className="SignUp">
+            <div className="SignUp__top">
+              <button className='SignUp__top-link' onClick={() => openModal('register')} >Зарегистрируйтесь</button>,<br/>
+              чтобы начать добавлять<br/>
+              рецепты
+            </div>
+            <img src="./img/girlMain.png" alt="" className='SignUp__girl'/>
+            <div className="SignUp__bottom"></div>
+          </div>
+      )}
     </div>
   );
 };
