+++ src/pages/recipe/Recipe.js
@@ -2,7 +2,6 @@
 import React from 'react';
 import { useParams, Link, useNavigate } from 'react-router-dom';
 import '../../styles/Recipe.css';
-import Header from '../../components/Header';
 import Footer from '../../components/Footer';
 
 const Recipe = () => {
@@ -65,7 +64,6 @@ const Recipe = () => {
     };
     return (
         <div className='Recipe'>
-            <Header img={recipe.logo} />
             <h1 className='Recipe__title'>{recipe.name}</h1>
             <div className="Recipe__top">
             <img src={recipe.img} alt="" className='Recipe__top-img' />
