+++ src/styles/RecipesMain.css
@@ -112,6 +112,13 @@
   padding-top: 50px;
 }
 
+.SignUp__topBlank {
+  width: 330px;
+  height: 169px;
+  border-radius: 30px;
+  background-color: transparent;
+}
+
 .SignUp__top-link {
   color: rgba(206, 32, 32, 1);
   font-weight: 900;
